Desktop/BLOC_3/Docker/requirements.txt
pandas
psycopg2-binary
python-dotenv
prometheus_client
